+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 * *[v0.2.0-rc4]*
   * [FIX]: Fix for an issue that caused YouTube videos set to autoplay
     to load during animations
+  * [DEV]: Add ability to see and debug uncompressed ES2015 code in
+    production
 
 * *[v0.2.0-rc3]*
   * [FIX]: Fix for an issue that caused an advertiser's name not to be